+++ AmplifyPlugins/DataStore/AWSDataStoreCategoryPluginTests/Sync/RemoteSyncEngineTests.swift
@@ -68,6 +68,7 @@ class RemoteSyncEngineTests: XCTestCase {
         let mutationsPaused = expectation(description: "mutationsPaused")
         let stateMutationsCleared = expectation(description: "stateMutationsCleared")
         let subscriptionsInitialized = expectation(description: "subscriptionsInitialized")
+        let subscriptionsEstablishedReceived = expectation(description: "subscriptionsEstablished received")
         let cleanedup = expectation(description: "cleanedup")
         let failureOnInitialSync = expectation(description: "failureOnInitialSync")
 
@@ -75,6 +76,17 @@ class RemoteSyncEngineTests: XCTestCase {
 
         let advice = RequestRetryAdvice.init(shouldRetry: false)
         mockRequestRetryablePolicy.pushOnRetryRequestAdvice(response: advice)
+
+        let filter = HubFilters.forEventName(HubPayload.EventName.DataStore.subscriptionsEstablished)
+        let hubListener = Amplify.Hub.listen(to: .dataStore, isIncluded: filter) { payload in
+            XCTAssertNil(payload.data)
+            subscriptionsEstablishedReceived.fulfill()
+        }
+
+        guard try HubListenerTestUtilities.waitForListener(with: hubListener, timeout: 5.0) else {
+            XCTFail("Listener not registered for hub")
+            return
+        }
 
         let remoteSyncEngineSink = remoteSyncEngine
             .publisher
@@ -113,9 +125,11 @@ class RemoteSyncEngineTests: XCTestCase {
                    mutationsPaused,
                    stateMutationsCleared,
                    subscriptionsInitialized,
+                   subscriptionsEstablishedReceived,
                    cleanedup,
                    failureOnInitialSync], timeout: defaultAsyncWaitTimeout)
         remoteSyncEngineSink.cancel()
+        Amplify.Hub.removeListener(hubListener)
     }
 
     func testRemoteSyncEngineHappyPath() throws {